hwmon: ina230: add type casting for negative current
When precision_multiplier has positive value and current has negative
value, the calculation makes wrong result by implicit type casting.

Bug 1311630

Reviewed-on: http://git-master/r/240533
(cherry picked from commit cb432c7eac648a11602f523fabf05c9662856e7b)
Reviewed-on: http://git-master/r/244215
GVS: Gerrit_Virtual_Submit
Tested-by: Hyong Bin Kim <hyongbink@nvidia.com>

+++ drivers/hwmon/ina230.c
@@ -461,7 +461,7 @@ static s32 show_current(struct device *dev,
 	if (data->pdata->divisor)
 		current_mA /= (s16) data->pdata->divisor;
 	if (data->pdata->precision_multiplier)
-		current_mA /= data->pdata->precision_multiplier;
+		current_mA /= (s16) data->pdata->precision_multiplier;
 
 	return sprintf(buf, "%d mA\n", current_mA);
 }
